implement publisher confirms
there are occasion when one would like to wait for [publisher
confirms](https://www.rabbitmq.com/confirms.html). one use case is
when running a kubernetes cron job that publishes messages to rabbitmq
and dies. sometimes, the erlang vm shuts down before the message is
fully transfered to the server, resulting in data loss.

this introduces options to use amqp's publisher confirms api.

to use, one should pass the following options to `ConduitAMQP.publish/4`:

- `:publisher_confirms` :: should be one of `:no_confirmation`,
`:wait` or `:die`. defaults to `:no_confirmation`
- `:publisher_confirms_timeout` :: if there should be a timeout in
milliseconds for the server acknowledgement. should be `:infinity` for
no timeout or an integer in milliseconds.